Extract songs index core
